+++ core/execution/future/market.go
@@ -149,6 +149,7 @@ type Market struct {
 	expiringStopOrders      *common.ExpiringOrders
 
 	minDuration time.Duration
+	perp        bool
 }
 
 // NewMarket creates a new market using the market framework configuration and creates underlying engines.
@@ -257,6 +258,7 @@ func NewMarket(
 	}
 
 	mkt.MarketTimestamps = ts
+	isPerp := mkt.MarketType() == types.MarketTypePerp
 
 	market := &Market{
 		log:                       log,
@@ -297,14 +299,19 @@ func NewMarket(
 		maxStopOrdersPerParties:   num.UintZero(),
 		stopOrders:                stoporders.New(log),
 		expiringStopOrders:        common.NewExpiringOrders(),
+		perp:                      isPerp,
 	}
 
 	assets, _ := mkt.GetAssets()
 	market.settlementAsset = assets[0]
 
 	liqEngine.SetGetStaticPricesFunc(market.getBestStaticPricesDecimal)
-	market.tradableInstrument.Instrument.Product.NotifyOnTradingTerminated(market.tradingTerminated)
-	market.tradableInstrument.Instrument.Product.NotifyOnSettlementData(market.settlementData)
+	if !isPerp {
+		market.tradableInstrument.Instrument.Product.NotifyOnTradingTerminated(market.tradingTerminated)
+		market.tradableInstrument.Instrument.Product.NotifyOnSettlementData(market.settlementData)
+	} else {
+		market.tradableInstrument.Instrument.Product.NotifyOnSettlementData(market.settlementDataPerp)
+	}
 	market.assetDP = uint32(assetDetails.DecimalPlaces())
 	return market, nil
 }
@@ -386,12 +393,16 @@ func (m *Market) Update(ctx context.Context, config *types.Market, oracleEngine 
 	m.liquidity.UpdateMarketConfig(m.tradableInstrument.RiskModel, m.pMonitor)
 
 	// if we're already in trading terminated, not point to listen to trading termination oracle
-	if m.mkt.State != types.MarketStateTradingTerminated {
+	if m.perp {
+		m.tradableInstrument.Instrument.Product.NotifyOnSettlementData(m.settlementDataPerp)
+	} else if mkt.State != types.MarketStateTradingTerminated {
 		m.tradableInstrument.Instrument.Product.NotifyOnTradingTerminated(m.tradingTerminated)
 	} else {
 		m.tradableInstrument.Instrument.UnsubscribeTradingTerminated(ctx)
 	}
-	m.tradableInstrument.Instrument.Product.NotifyOnSettlementData(m.settlementData)
+	if !m.perp {
+		m.tradableInstrument.Instrument.Product.NotifyOnSettlementData(m.settlementData)
+	}
 
 	m.updateLiquidityFee(ctx)
 	// risk model hasn't changed -> return
@@ -2016,7 +2027,6 @@ func (m *Market) confirmMTM(
 	mp := m.getCurrentMarkPrice()
 	evts := m.position.UpdateMarkPrice(mp)
 	settle := m.settlement.SettleMTM(ctx, mp, evts)
-	orderUpdates := []*types.Order{}
 
 	// let the product know about the mark-price, incase its the sort of product that cares
 	if m.markPrice != nil {
@@ -2025,56 +2035,7 @@ func (m *Market) confirmMTM(
 
 	// Only process collateral and risk once per order, not for every trade
 	margins := m.collateralAndRisk(ctx, settle)
-	if len(margins) > 0 {
-		transfers, closed, bondPenalties, err := m.collateral.MarginUpdate(ctx, m.GetID(), margins)
-		if err != nil {
-			m.log.Error("margin update had issues", logging.Error(err))
-		}
-		if err == nil && len(transfers) > 0 {
-			evt := events.NewLedgerMovements(ctx, transfers)
-			m.broker.Send(evt)
-		}
-		if len(bondPenalties) > 0 {
-			transfers, err := m.bondSlashing(ctx, bondPenalties...)
-			if err != nil {
-				m.log.Error("Failed to perform bond slashing",
-					logging.Error(err))
-			}
-			// if bond slashing occurred then amounts in "closed" will not be accurate
-			if len(transfers) > 0 {
-				m.broker.Send(events.NewLedgerMovements(ctx, transfers))
-				closedRecalculated := make([]events.Margin, 0, len(closed))
-				for _, c := range closed {
-					if pos, ok := m.position.GetPositionByPartyID(c.Party()); ok {
-						margin, err := m.collateral.GetPartyMargin(pos, m.settlementAsset, m.mkt.ID)
-						if err != nil {
-							m.log.Error("couldn't get party margin",
-								logging.PartyID(c.Party()),
-								logging.Error(err))
-							// keep old value if we weren't able to recalculate
-							closedRecalculated = append(closedRecalculated, c)
-							continue
-						}
-						closedRecalculated = append(closedRecalculated, margin)
-					}
-				}
-				closed = closedRecalculated
-			}
-		}
-		if len(closed) > 0 {
-			upd, err := m.resolveClosedOutParties(
-				ctx, closed)
-			if err != nil {
-				m.log.Error("unable to closed out parties",
-					logging.String("market-id", m.GetID()),
-					logging.Error(err))
-			}
-			if len(upd) > 0 {
-				orderUpdates = append(orderUpdates, upd...)
-			}
-		}
-		m.updateLiquidityFee(ctx)
-	}
+	orderUpdates := m.handleRiskEvts(ctx, margins)
 
 	// orders updated -> check reference moves
 	// force check
@@ -2084,6 +2045,62 @@ func (m *Market) confirmMTM(
 		// release excess margin for all positions
 		m.recheckMargin(ctx, m.position.Positions())
 	}
+}
+
+func (m *Market) handleRiskEvts(ctx context.Context, margins []events.Risk) []*types.Order {
+	if len(margins) == 0 {
+		return
+	}
+	orderUpdates := []*types.Order{}
+	transfers, closed, bondPenalties, err := m.collateral.MarginUpdate(ctx, m.GetID(), margins)
+	if err != nil {
+		m.log.Error("margin update had issues", logging.Error(err))
+	}
+	if err == nil && len(transfers) > 0 {
+		evt := events.NewLedgerMovements(ctx, transfers)
+		m.broker.Send(evt)
+	}
+	if len(bondPenalties) > 0 {
+		transfers, err := m.bondSlashing(ctx, bondPenalties...)
+		if err != nil {
+			m.log.Error("Failed to perform bond slashing",
+				logging.Error(err))
+		}
+		// if bond slashing occurred then amounts in "closed" will not be accurate
+		if len(transfers) > 0 {
+			m.broker.Send(events.NewLedgerMovements(ctx, transfers))
+			closedRecalculated := make([]events.Margin, 0, len(closed))
+			for _, c := range closed {
+				if pos, ok := m.position.GetPositionByPartyID(c.Party()); ok {
+					margin, err := m.collateral.GetPartyMargin(pos, m.settlementAsset, m.mkt.ID)
+					if err != nil {
+						m.log.Error("couldn't get party margin",
+							logging.PartyID(c.Party()),
+							logging.Error(err))
+						// keep old value if we weren't able to recalculate
+						closedRecalculated = append(closedRecalculated, c)
+						continue
+					}
+					closedRecalculated = append(closedRecalculated, margin)
+				}
+			}
+			closed = closedRecalculated
+		}
+	}
+	if len(closed) > 0 {
+		upd, err := m.resolveClosedOutParties(
+			ctx, closed)
+		if err != nil {
+			m.log.Error("unable to closed out parties",
+				logging.String("market-id", m.GetID()),
+				logging.Error(err))
+		}
+		if len(upd) > 0 {
+			orderUpdates = append(orderUpdates, upd...)
+		}
+	}
+	m.updateLiquidityFee(ctx)
+	return orderUpdates
 }
 
 // updateLiquidityFee computes the current LiquidityProvision fee and updates
@@ -3714,6 +3731,49 @@ func (m *Market) settlementData(ctx context.Context, settlementData *num.Numeric
 		return
 	}
 	m.settlementDataWithLock(ctx, types.MarketStateSettled, settlementDataInAsset)
+}
+
+func (m *Market) settlementDataPerp(ctx context.Context, settlementData *num.Numeric) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	m.settlementDataInMarket = settlementData
+	settlementDataInAsset, err := m.tradableInstrument.Instrument.Product.ScaleSettlementDataToDecimalPlaces(m.settlementDataInMarket, m.assetDP)
+	if err != nil {
+		m.log.Error(err.Error())
+		return
+	}
+	// take all positions, get funding transfers
+	transfers, round := m.settlement.SettleFundingPeriod(t, m.position.Positions(), settlementPriceInAsset)
+	if len(positions) == 0 {
+		m.log.Debug("Failed to get settle positions for funding period")
+		return
+	}
+
+	margins, ledgerMovements, err := m.collateral.PerpsFundingSettlement(ctx, m.GetID(), positions, round)
+	if err != nil {
+		m.log.Error("Failed to get ledger movements when performing the funding settlement",
+			logging.MarketID(m.GetID()),
+			logging.Error(err))
+		return
+	}
+
+	if len(ledgerMovements) > 0 {
+		m.broker.Send(events.NewLedgerMovements(ctx, ledgerMovements))
+	}
+	// final distribution of liquidity fees
+	m.distributeLiquidityFees(ctx)
+	// no margin events, no margin stuff to check
+	if len(margins) == 0 {
+		return
+	}
+
+	riskUpdates := m.risk.UpdateMarginsOnSettlement(ctx, margins, m.getCurrentMarkPrice())
+	if len(riskUpdates) == 0 {
+		return
+	}
+	orderUpdates := m.handleRiskEvts(ctx, riskUpdates)
+	m.checkForReferenceMoves(
+		ctx, orderUpdates, false)
 }
 
 // NB this must be called with the lock already acquired.