+++ core/types/market.go
@@ -887,6 +887,14 @@ func (m MarketData) String() string {
 	)
 }
 
+type MarketType uint32
+
+const (
+	MarketTypeFuture MarketType = iota
+	MarketTypeSpot
+	MarketTypePerp
+)
+
 type Market struct {
 	ID                            string
 	TradableInstrument            *TradableInstrument
@@ -906,6 +914,7 @@ type Market struct {
 	MarketTimestamps      *MarketTimestamps
 	ParentMarketID        string
 	InsurancePoolFraction num.Decimal
+	mType                 MarketType
 }
 
 func MarketFromProto(mkt *vegapb.Market) (*Market, error) {
@@ -942,6 +951,12 @@ func MarketFromProto(mkt *vegapb.Market) (*Market, error) {
 		QuadraticSlippageFactor:       quadraticSlippageFactor,
 		ParentMarketID:                parent,
 		InsurancePoolFraction:         insFraction,
+		mType:                         MarketTypeFuture,
+	}
+	if s := m.GetSpot(); s != nil {
+		m.mType = MarketTypeSpot
+	} else if p := m.GetPerps(); p != nil {
+		m.mType = MarketTypePerp
 	}
 
 	if mkt.LiquiditySlaParams != nil {
@@ -1031,6 +1046,10 @@ func (m Market) String() string {
 		m.State.String(),
 		stringer.ReflectPointerToString(m.MarketTimestamps),
 	)
+}
+
+func (m Market) MarketType() MarketType {
+	return m.mType
 }
 
 func (m Market) DeepClone() *Market {